refactor player to extend moveableEntity and added new moveable entity bullet
