+++ src/com/sgrostad/entities/Entity.java
@@ -41,6 +41,7 @@ public abstract class Entity {
 
     public boolean checkEntityCollision(float xOffset, float yOffset){
         for(Entity e : handler.getWorld().getEntityManager().getEntities()){
+            //TODO Have experienced an error here, but have not managed to reproduce it. Happened when firing many bullets.
             if (e.equals(this))
                 continue;
             if (e.getCollisionBounds(0f, 0f).intersects(getCollisionBounds(xOffset, yOffset))){
