+++ src/com/sgrostad/input/player/PlayerTakeOffTimer.java
@@ -1,6 +1,6 @@
 package com.sgrostad.input.player;
 
-import com.sgrostad.entities.creatures.Player;
+import com.sgrostad.entities.mobiles.Player;
 
 public class PlayerTakeOffTimer {
     private final float STANDARD_MINIMUM_JUMP_PERCENTAGE = 0.7f;
